refactor(040): migra packages/design-tokens para TypeScript
